backend/requirements.txt
# FastAPI and web framework dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Database dependencies
supabase>=2.0.0
python-dotenv>=1.0.0
nanoid>=2.0.0
Pillow>=10.0.0

# ML dependencies
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
xgboost>=1.6.0
lightgbm>=3.3.0
catboost>=1.1.0
imbalanced-learn>=0.9.0
joblib>=1.1.0

# Data visualization (optional for deployment)
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Database and caching
aiofiles>=23.0.0

# Email functionality (for waitlist verification)
aiosmtplib>=3.0.0
email-validator>=2.0.0

# Environment and configuration
python-dotenv>=1.0.0

# Web scraping dependencies (if needed)
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0

# Async support
httpx>=0.25.0
openai>=1.93.0

# Background jobs / queue
celery>=5.3.0
redis>=5.0.0

# Billing
stripe>=11.0.0
